chore: add automated linting and formatting
Added automated linting and formatting of all project files before any
commit can succeed. A husky pre-commit hook was added to accomplish
this functionality with the lint-staged utility. JavaScript-type files
are linted with eslint, CSS files are linted with stylelint, and
everything is formatted with prettier. Additional scripts for manual
linting and formatting were also added to the package.json file.

All configuration related files are now placed in the "config" folder.